+++ src/main/java/action/CommandManager.java
@@ -1,4 +1,17 @@
+package action;
+
 import java.util.HashMap;
+
+import action.commands.Command;
+import action.commands.DeadlineCommand;
+import action.commands.DeleteCommand;
+import action.commands.EndCommand;
+import action.commands.EventCommand;
+import action.commands.ListCommand;
+import action.commands.MarkCommand;
+import action.commands.TodoCommand;
+import data.exception.BarneyException;
+import data.exception.InvalidCommandException;
 
 public class CommandManager {
