+++ src/main/java/action/commands/Command.java
@@ -1,13 +1,21 @@
+package action.commands;
+
 import java.util.HashMap;
 
-abstract class Command {
+import action.CommandManager;
+import data.TaskList;
+import data.exception.InvalidArgumentException;
+import data.exception.MissingFlagException;
+import ui.Ui;
+
+public abstract class Command {
     private final String name;
 
     public Command(String name) {
         this.name = name;
     }
 
-    abstract boolean execute(TaskList tasks, Ui ui) throws InvalidArgumentException;
+    public abstract boolean execute(TaskList tasks, Ui ui) throws InvalidArgumentException;
 
     void verifyFlags(HashMap<String, String> argumentMap) throws InvalidArgumentException {
         for (String arg : CommandManager.CommandType.fromString(this.name).commandArgs) {
