+++ src/main/java/action/commands/DeadlineCommand.java
@@ -1,4 +1,11 @@
+package action.commands;
+
 import java.util.HashMap;
+
+import data.TaskList;
+import data.exception.InvalidArgumentException;
+import data.task.DeadlineTask;
+import ui.Ui;
 
 public class DeadlineCommand extends Command {
     HashMap<String, String> argumentMap;
