+++ src/main/java/data/exception/BarneyException.java
@@ -1,3 +1,5 @@
+package data.exception;
+
 public class BarneyException extends Exception {
     public BarneyException(String message) {
         super(message);
