+++ src/main/java/data/exception/InvalidArgumentException.java
@@ -1,3 +1,5 @@
+package data.exception;
+
 public class InvalidArgumentException extends BarneyException {
     public InvalidArgumentException(String message) {
         super(message);
