+++ src/main/java/data/exception/InvalidCommandException.java
@@ -1,3 +1,5 @@
+package data.exception;
+
 public class InvalidCommandException extends BarneyException {
     public InvalidCommandException(String message) {
         super(message);
