+++ src/main/java/data/exception/InvalidSaveFormatException.java
@@ -1,3 +1,5 @@
+package data.exception;
+
 public class InvalidSaveFormatException extends BarneyException {
     public InvalidSaveFormatException(String message) {
         super("Invalid Save File Format: " + message);
