+++ src/main/java/data/exception/MissingFlagException.java
@@ -1,3 +1,5 @@
+package data.exception;
+
 public class MissingFlagException extends InvalidArgumentException {
     public MissingFlagException(String message) {
         super(message);
