+++ src/main/java/data/task/Task.java
@@ -1,3 +1,5 @@
+package data.task;
+
 import java.util.ArrayList;
 import java.util.Arrays;
