+++ src/main/java/ui/Ui.java
@@ -1,3 +1,8 @@
+package ui;
+
+import data.TaskList;
+import data.task.Task;
+
 public class Ui {
 
     public Ui() {
